backend/requirements.txt
flask
flask-cors
langchain
langchain-google-genai
langchain-community
langgraph
faiss-cpu
pymupdf
python-dotenv